requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
asyncpg==0.29.0
redis==5.0.1
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiohttp==3.9.1 